Refactoring store_ip_data logic
A visitor record will be created, as long as the ip address is routable.
If a request for data against ipinfo.io, returns data or not, this will
be conditionally processed, trying to obtain available data for country
and city. If the visitor (ip address) exists and the last_request date
is different than current date (yesterday, for example).