Redefining Middleware and set cookie behaviour
Conditions and processing of data were improved on the middleware,
moving the cookie expire timestamp calculation to set_cookie method.
Documentation was also added, explaining the big picture of the
Middleware class, as well as the set_cookie method workflow.
Also, the file and classname were redefined for the Middleware
structure.