feat: middleware refactor + long cookie expiration
Middelware name was changed, so as the function from geotracking.utils
module.

The function which actually registers a Visitor, was improved in terms
of logic and documentation.

And 'device_id' cookie has an expiration date for 5 years in the future,
and not until midnight: we want cookie persistency, in order to track
visitors.